+++ chess/src/main/java/pdp/model/ai/heuristics/MaterialHeuristic.java
@@ -7,12 +7,12 @@ import pdp.model.board.Board;
 /** Heuristic based on the number of pieces still on the board. */
 public class MaterialHeuristic implements Heuristic {
 
-  private static final int SCORE_CAP = 100;
-  private static final int PAWN_VALUE = 1;
-  private static final int QUEEN_VALUE = 9;
-  private static final int BISHOP_VALUE = 3;
-  private static final int KNIGHT_VALUE = 3;
-  private static final int ROOK_VALUE = 5;
+  private static final float SCORE_CAP = 100f;
+  private static final float PAWN_VALUE = 1f;
+  private static final float QUEEN_VALUE = 9f;
+  private static final float BISHOP_VALUE = 3f;
+  private static final float KNIGHT_VALUE = 3f;
+  private static final float ROOK_VALUE = 5f;
   private static final float MULTIPLIER =
       SCORE_CAP / (QUEEN_VALUE * 8 + BISHOP_VALUE * 2 + KNIGHT_VALUE * 2 + ROOK_VALUE * 2);
 
@@ -28,7 +28,7 @@ public class MaterialHeuristic implements Heuristic {
     if (!(board.getBoardRep() instanceof BitboardRepresentation bitboardRep)) {
       throw new InvalidBoardException();
     }
-    int score = 0;
+    float score = 0;
     score += (bitboardRep.getPawns(true).size() - bitboardRep.getPawns(false).size()) * PAWN_VALUE;
     score +=
         (bitboardRep.getQueens(true).size() - bitboardRep.getQueens(false).size()) * QUEEN_VALUE;
